Split autoroll_lts.py main and chromium logic.
Creates an autoroll_lib for shared logic between autoroll and
autoroll_chromium. Adds autoroll_chromium that batches cherry-picks and
is specific to chromium autorolls / assumes that every cherry-pick will
result in a conflict.

Bug: 543555380
